Add TXListPayload logic.
