+++ src/DBFTPlugin/Messages/TXListRequest.cs
@@ -4,12 +4,12 @@ using System.IO;
 
 namespace Neo.Consensus.Messages
 {
-    public class TXHashesRequest : ConsensusMessage
+    public class TXListRequest : ConsensusMessage
     {
 
         public override int Size => base.Size;
 
-        public TXHashesRequest() : base(ConsensusMessageType.TXHashesRequest) { }
+        public TXListRequest() : base(ConsensusMessageType.TXListRequest) { }
 
         public override void Deserialize(BinaryReader reader)
         {
