+++ src/DBFTPlugin/Messages/TXListResponse.cs
@@ -5,21 +5,18 @@ using System.Linq;
 
 namespace Neo.Consensus.Messages
 {
-    public class TXHashesResponce : ConsensusMessage
+    public class TXListResponse : ConsensusMessage
     {
-        public byte[] Signature;
         public UInt256[] TransactionHashes;
 
         public override int Size => base.Size
-            + Signature.Length                  
             + TransactionHashes.GetVarSize();   //TransactionHashes;
 
-        public TXHashesResponce() : base(ConsensusMessageType.TXHashesResponce) { }
+        public TXListResponse() : base(ConsensusMessageType.TXListResponse) { }
 
         public override void Deserialize(BinaryReader reader)
         {
             base.Deserialize(reader);
-            Signature = reader.ReadFixedBytes(64);
             TransactionHashes = reader.ReadSerializableArray<UInt256>(ushort.MaxValue);
             if (TransactionHashes.Distinct().Count() != TransactionHashes.Length)
                 throw new FormatException();
@@ -28,7 +25,6 @@ namespace Neo.Consensus.Messages
         public override void Serialize(BinaryWriter writer)
         {
             base.Serialize(writer);
-            writer.Write(Signature);
             writer.Write(TransactionHashes);
         }
     }
